fix: show new videos counter on library items
